requirements.txt
google-cloud-firestore==1.6.0
grpcio==1.26.0
